+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/BridgeInternal.java
@@ -143,16 +143,14 @@ public final class BridgeInternal {
         QueryInfo.QueryPlanDiagnosticsContext diagnosticsContext,
         boolean useEtagAsContinuation,
         boolean isNoChangesResponse,
-        CosmosDiagnostics cosmosDiagnostics,
-        Boolean hasMoreChangesToProcess) {
+        CosmosDiagnostics cosmosDiagnostics) {
         FeedResponse<T> feedResponseWithQueryMetrics = ModelBridgeInternal.createFeedResponseWithQueryMetrics(
             results,
             headers,
             queryMetricsMap,
             diagnosticsContext,
             useEtagAsContinuation,
-            isNoChangesResponse,
-            hasMoreChangesToProcess);
+            isNoChangesResponse);
 
         ClientSideRequestStatistics requestStatistics;
         if (cosmosDiagnostics != null) {
@@ -178,25 +176,6 @@ public final class BridgeInternal {
         }
 
         return feedResponseWithQueryMetrics;
-    }
-
-    @Warning(value = INTERNAL_USE_ONLY_WARNING)
-    public static <T> FeedResponse<T> createFeedResponseWithQueryMetrics(
-        List<T> results,
-        Map<String, String> headers,
-        ConcurrentMap<String, QueryMetrics> queryMetricsMap,
-        QueryInfo.QueryPlanDiagnosticsContext diagnosticsContext,
-        boolean useEtagAsContinuation,
-        boolean isNoChangesResponse,
-        CosmosDiagnostics cosmosDiagnostics) {
-        return createFeedResponseWithQueryMetrics(results,
-            headers,
-            queryMetricsMap,
-            diagnosticsContext,
-            useEtagAsContinuation,
-            isNoChangesResponse,
-            cosmosDiagnostics,
-            null);
     }
 
     @Warning(value = INTERNAL_USE_ONLY_WARNING)