+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/CosmosAsyncContainer.java
@@ -1187,8 +1187,6 @@ public class CosmosAsyncContainer {
         boolean isNoChangesResponse = isChangeFeed ?
             ModelBridgeInternal.getNoChangesFromFeedResponse(response)
             : false;
-        Boolean hasMoreChangesToProcess = ImplementationBridgeHelpers.FeedResponseHelper.getFeedResponseAccessor()
-                .getHasMoreChangesToProcess(response);
 
         return BridgeInternal.createFeedResponseWithQueryMetrics(
             response.getResults(),
@@ -1197,8 +1195,7 @@ public class CosmosAsyncContainer {
             ModelBridgeInternal.getQueryPlanDiagnosticsContext(response),
             useEtagAsContinuation,
             isNoChangesResponse,
-            response.getCosmosDiagnostics(),
-            hasMoreChangesToProcess);
+            response.getCosmosDiagnostics());
     }
 
     /**