+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/feedranges/FeedRangeContinuation.java
@@ -139,6 +139,8 @@ public abstract class FeedRangeContinuation extends JsonSerializable {
     public abstract <T> boolean hasFetchedAllChanges(
         FeedResponse<T> responseMessage, Long endLSN);
 
+    public abstract boolean hasFetchedAllChanges();
+
     public abstract Mono<ShouldRetryResult> handleFeedRangeGone(
         RxDocumentClientImpl client,
         GoneException goneException);